+++ photons_app/task_finder.py
@@ -5,19 +5,13 @@ Responsible for finding tasks and executing them
 from photons_app.actions import available_actions, all_tasks
 from photons_app.errors import BadTask
 
-from delfick_project.norms import sb
-
 
 class TaskFinder(object):
     def __init__(self, collector):
         self.tasks = all_tasks
         self.collector = collector
 
-    async def task_runner(self, task, **kwargs):
-        target = sb.NotSpecified
-        if ":" in task:
-            target, task = task.split(":", 1)
-
+    async def task_runner(self, target, task, **kwargs):
         if task not in self.tasks:
             raise BadTask("Unknown task", task=task, available=sorted(list(self.tasks.keys())))